AI-Bot/requirements.txt
openai-whisper
transformers
langchain
langchain-huggingface
huggingface_hub
sounddevice
soundfile
numpy
requests
keyboard
python-dotenv
sounddevice
python-dotenv
langchain
langchain-groq
edge-tts
pygame
numpy


# Download FFmpeg